Moved certain ride management functionality to utilities.
