+++ EyeCT4Rails/EyeCT4RailsLogic/Repositories/RideManagementRepository.cs
@@ -1,12 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using EyeCT4RailsDatabase.Models;
 using EyeCT4RailsDatabase.SQLContexts;
-using EyeCT4RailsLib;
-using EyeCT4RailsLib.Classes;
 using EyeCT4RailsLib.Enums;
 using EyeCT4RailsLogic.Exceptions;
+using EyeCT4RailsLogic.Utilities;
 
 // ReSharper disable MemberCanBeMadeStatic.Local
 
@@ -40,91 +37,9 @@ namespace EyeCT4RailsLogic.Repositories
             }
             catch (Exception e)
             {
-                LogicExceptionHandler.FilterOracleDatabaseException(e);
+                ExceptionHandler.FilterOracleDatabaseException(e);
                 throw new UnknownException("FATAL ERROR! EXTERMINATE! EXTERMINATE!");
             }
-        }
-
-        /// <summary>
-        /// Gets a free section from a given depot.
-        /// </summary>
-        /// <param name="depot">The depot in question.</param>
-        /// <returns>The first free section it finds.</returns>
-        public Section GetFreeSection(Depot depot, TramType type)
-        {
-            var tracks = depot.Tracks;
-
-            switch (type)
-            {
-                case TramType.DCombino:
-                case TramType.Trainer:
-                    var track57 = tracks.Find(t => t.Id == 57);
-                    tracks.Remove(track57);
-                    tracks.Add(track57);
-                    break;
-            }
-
-            var track40 = tracks.Find(t => t.Id == 40);
-            tracks.Remove(track40);
-            tracks.Add(track40);
-
-            var track58 = tracks.Find(t => t.Id == 58);
-            tracks.Remove(track58);
-            tracks.Add(track58);
-
-
-            foreach (Section ret in tracks.Select(GetFreeSection).Where(ret => ret != null))
-            {
-                return ret;
-            }
-
-            throw new NoFreeSectionException("No section is free");
-
-        }
-
-        /// <summary>
-        /// Loops through all sections in a track to get a free section.
-        /// </summary>
-        /// <param name="track">The track in question.</param>
-        /// <returns>A free section.</returns>
-        private Section GetFreeSection(Track track)
-        {
-            //Fetches all sections that have access to the outside world.
-            List<Section> freeSections =
-                track.Sections.Where(
-                    section => CheckSectionFreedom(section, false) || CheckSectionFreedom(section, true)).ToList();
-
-            if (freeSections.Count == 0)
-                return null;
-
-            //Tries to get a section next to an already blocked/taken section.
-            var ret =
-                freeSections.Find(x => (x.NextSection != null && (x.NextSection.Blocked || x.NextSection.Tram != null))
-                                       ||
-                                       (x.PreviousSection != null &&
-                                        (x.PreviousSection.Blocked || x.PreviousSection.Tram != null)));
-
-            //If the whole track is empty, i.e. ret is null then it will place the tram on the last section of the track.
-            return ret ?? freeSections.Last();
-        }
-
-        /// <summary>
-        /// Checks wheter a section is accessible from the outside.
-        /// </summary>
-        /// <param name="section">The section to check for.</param>
-        /// <param name="direction">The direction in which to look. True is next.</param>
-        /// <returns>A bool that is true, if and only if it can reach the outside.</returns>
-        public bool CheckSectionFreedom(Section section, bool direction)
-        {
-            if (section == null)
-                return true;
-            if (section.Blocked || section.Tram != null)
-                return false;
-            if (section.NextSection == null || section.PreviousSection == null)
-                return true;
-            return direction
-                ? CheckSectionFreedom(section.NextSection, true)
-                : CheckSectionFreedom(section.PreviousSection, false);
-        }
+        }      
     }
 }
